Move Buffer-related C++ files to node/
Move BufferEncodingType, JSBuffer, JSBufferEncodingType, JSBufferList,
JSStringDecoder from bindings/ to node/.
Updated BuildBun.cmake and NodeBufferModule.h, NodeStringDecoderModule.h.
